reverted visual test
